clean up structparse and a few imports
